requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15
dj-database-url>=2.1
psycopg2-binary>=2.9
django-import-export>=3.3
qrcode[pil]>=7.4
python-decouple>=3.8
whitenoise>=6.6
